node: Improve crate organization
